mk/su-make-errors.sh: allow cross-compile ((Jörg Schilling,) Roberto Ricci)..
Roberto wrote to the ML

  I'm trying to cross-compile s-nail version 14.9.19 on a Void Linux host.

  I run the command:
      make CC=name_of_cross_compiler OPT_CROSS_BUILD=yes config
  but I get the following error:
      [...]
       . OS error mapping table generated ... no
      make: *** [makefile:23: config] Error 1

  This probably happens because the config script needs to compile a program
  and then run it. Obviously it fails if I am cross-compiling.
  Normal compilation has no trouble.

So i totally missed this, what a mess :-|

Rewrite this script to use the C preprocessor to resolve the OS
error numbers, but perform all the mapping etc. stuff via AWK.

While doing this i remembered that Jörg pointed me into the
direction a few years back, but i was not listening.
Even more thanks to Roberto for being crystal clear. :)

  Developing this revealed a SEGV in nawk and a faulty
  parser/expander of busybox awk (expanded
    osnoa[i] = "(su__ERR_NUMBER_MAX - " ++unavail ")"
  to 00) etc, workaround was
    ++unavail
    osnoa[i] = "(su__ERR_NUMBER_MAX - " unavail ")"
  Both problems have been reported).