n_go_main_loop(): fix: do not loose exit status with *on-history-addition*..
We falsely set the actual desire / necessity to exit the main loop
after calling a possibly set *on-history-addition*, which is wrong
because a_go_evaluate() will reset n_exit_status in that case
(after the first non-erroneous command of that hook).  The
(*errexit* caused in this case) "fault" error status would be used
to cause an exit nonetheless, which is why this error lingered.