[net-test] Fix EXTERNAL/EXTERNANON shadowing..
a_netsmtp_parse_caps() for example calls cred_auth_type_find_name()
to parse server AUTH capabilities, then sets the according bit.
Like so EXTERNANON turned out to be untestable since the announced
EXTERNAL would be mapped to EXTERNAL only.

Explicitly document this hack, which it is, and ensure callees get
this special condition right.  It is of no value to introduce
a generalization like cred_auth_type_extend_to_mask() or whatever,
this EXTERNAL/EXTERNANON thing is owed due to our own constraint
that there shall be only one *user*, which, if set, is also used
in credential handshakes, and the problem that some servers do not
actually truly support EXTERNAL, even if the name would match, for
example (the tested versions of) postfix only accept the name from
the certificate, and nothing else.