FIXME [-] (BWDIC!) Rework Reply-To:, fix *-add-cc* <-> Reply-To:, etc. (Geoff Clare)..
After a discussion on the POSIX list i realized that our Reply-To:
handling is not forbidden but the most strange possible,
effectively acting like the never standardized Mail-Followup-To:.
Not only because i stumbled about the consequences myself multiple
times i think it is for the much better if this behaviour is
simply thrown overboard.

The email-standard desired behaviour requires two new pseudo
headers so that hooks etc. know about edited origins (and
sender), so invent Mailx-Edited-Sender: and Mailx-Edited-Origin:
pseudo header fields.

It also requires changes to our *forward-add-cc* and
*quote-add-cc* functionality, which now have to take into account
a possibly set Reply-To:.

Alongside all this try to organize the code a little bit more
sensible.  At least until all that gets hidden under an object
interface, shall we ever reach this state.