[-] mx-test.sh: finally take advantage of our per-test directories..
Convert almost all tests to not use $MBOX and the other global
constants which were used when all the data the test produced was
injected in the build directory (which even earlier was the normal
source directory), but instead to take advantage of the fact that
each and every test is now run in its own subdirectory: we can
leave files around and create as many as we want.

This has the nice side effect that $MAILX_CC_TEST_NO_CLEANUP
likely results in some useful files laying around.

While here, add vim foldmarks.

For "git rebase" simplicity on this development branch which is in
terrible state since i think 14 months (!) do not convert all, but
only those which do not clash with commits on the [net-test] topic
branch -- do the conversation there.

NOTE: due to rebasing this is a bit weird, and requires the
(send.c:_send_al7ive_have_better(): favour-rich was never but
a hack) changeset that comes later in history.