[-] MIME classification: support EXT chains (.tar.gz..) (Olaf Hering)..
It was a conscious decision in the past to not allow extension
hierarchies, because IANA does not know about them, and they are
not inter-operable.

So for example tar.gz is application/gzip and nothing else.  The
"else" will appear once this MIME type has been expanded, the real
solution would thus be a "recursion until no more expansion is
possible" approach.

But that is not possible via mailcap etc., and therefore (i
presume) people started using things like x-tar-gz, because now
they could define an entry that gives the desired end result
without any further action, causing grief on receiver systems
because of the unknown MIME type.

In an upcoming commit we will have to introduce "inbound MIME
types" in order to allow separation of classification for inbound
and outbound code paths.