mx_socket_reset_read_buf(): rbufptr=NIL to avoid bad copies (Jörg Schilling)..
Jörg reported a bad pointer dereference on a Solaris platform, and
could convince that it is a bug of mine: old code i want to
rewrite creates stack objects in order to replace single instance
objects via bitcopy once the setup is complete.  (Instead of
having a notion of real objects.)  It happens that some newer code
does not nicely together with that pragmatic old style, which
i did not see nor think about (even though i should have and could
have).