c_digmsg(): (fix): if `digmsg' input is not fully consumed..
Do not (likely) crash when 'digmsg remove' removes a digmsg object
that is opened on a message and that still has valid input to be
consumed -- in that case an overlay pointer is set that is
detected and used by go_input() (if nothing else overlays this, of
course, this all is a bit complicated and should be event based
and emit LineCompletedEvents etc., like stated a billion times).

The fix is of course to reset the overlay in this case.
Adjust the manual to not only show in example but also state that
all input must be consumed when asking a `digmsg' object
a question.