sendpart(): fix [b735b9b1db6980e040fc97f22fd94ed7ae547615] for RFC822 parts..
We really need the MIME layer rewrite, this recursive thing is
such a terrible one, and you will just never be able to get that
right.  So if a message contains nothing but just an inner RFC822
message, the mentioned commit broke stuff, meaning that it did not
honour *rfc822-body-from_* and printed the inner headers directly
following the ones of the outer envelope.

It has to be said: we do not have proper tests.  I do have several
hundred messages with the one or the other problem, but this is
not automatized.  (I found it by going through all of them.)

The problem is that all tests which go for output etc will have to
be adjusted once we can gracefully place part infos etc., and the
number of empty separating messages (now impossible!) in between
them etc etc.

I should have done this many years ago.