Impose upper limit on socket failures (Stephen Isard)..
He reported endless blocks after connection breakup.
Indeed EAGAIN on blocking socket is not treated as hard error:

  IMAP ALARM 1
  NOOP 1
  IMAP FOR SOCKET WRITE
  NOOP 1 - out
  IMAP FOR ANSWER
  SOCKET TLS READ PRE AGAIN
  SOCKET TLS READ PRE
[SO_RCVTIMEO wait..]
  SOCKET TLS READ POST -1 error Resource temporarily unavailable
  SOCKET TLS READ PRE
  SOCKET TLS READ POST -1 error Resource temporarily unavailable
  SOCKET TLS READ PRE
  SOCKET TLS READ POST -1 error Resource temporarily unavailable
  ...

Just wondering why Linux reports this not EPIPE when the other
side of the road is really, really gone?
And i have not yet figured out what the total number of retries
is, it seems to last for about the quarter of an hour.
Since we really want RCVTIMEO, ... upper limit and complain:

  s-nail: TLS socket read error, retrying: Resource temporarily unavailable
  s-nail: TLS socket read error, retrying: Resource temporarily unavailable
  s-nail: IMAP: error:00000000:lib(0):func(0):reason(0)
  IMAP ALARM 2
  CHILD POST WAIT