[SU.2] SU: change avopt long identification approach..
It never has been a problem in the past, but now, with
avopt_parse_line(), and therefore the possibility to drive a very
simple command-line-option-based configuration file format through
avopt, being able to easily address more than 127
non-short-option-equivalents, and more than 256 long options
altogether, seems to be an important improvement.

(I think cURL and gpg2 may come close to this limit, or even
excess it as of today.  And whereas i am in favour of .ini file
syntax and an according object, being able to use avopt for that
very purpose for the s-postgrey that i work is a good thing (tm).)