Try fix/rewrite conditionals, I..
Our conditional code can easily be fooled, for example all the
following to not work as desired currently:

  ! -z [
  ! -z ]
  ! -z &&
  ! -z ||
  ! && == &&
  ! || == ||
  ! && != ||
  ! || != &&
  [ == [
  ! ! [ == [
  ! [ != [
  [ [ == [ ]
  [ ! ! [ == [ ]
  [ ! [ != [ ]
  [ [ [ [ == [ ] ] ]
  ! ! ] == ]
  ! ] != ]
  [ ! ! ] == ] ]
  [ ! ] != ] ]

For test(1) POSIX defined four explicit cases in order to be
tested, saying all bets are off for anything else.  But i like to
continue to use and-or lists and multiple tests per conditional!

This is a step that improves a bit but does not bring the healing
for these mentioned cases.  (I wanted to implement $^#,$^0,$^1
stuff now, and rebasing that was too difficult.)