SU dict: FIX: use correct pointer for reset_{insert,replace}()..
Added in [8898435070ada69f4ee80ca508763d593f250162] this functionality
was not part of the C++ implementation from which this is derived.
Unfortunately the added test was not sufficient to find the bug,
as it is dependent upon hash size and resize treshold.
A better test will be added on the development branch