mx_sendout_header_date(): support sub-minute precision (Arne Wörner)..
Arne reported on the list

  When I use this timezone file ([.]zoneinfo-TAI.tar.xz), and when
  I send an email with /bin/mail at 2022-07-28T04:48:54+00:00:37,
  then s-mailx writes: Date: Thu, 28 Jul 2022 04:48:54 +0000
  [.]Btw: That is what exim does (in my understanding), since
  i pointed it out...[.]

So what happens is that RFC 5322 and its predecessors do not
support timezones with sub-minute precision but via timezone
names, but which are more or less obsolete.  I started
a discussion [1], which became lengthy.  Regardless of that,
follow what exim did, and that is that it switches to the UTC
gmtime(3) in case the localetime(3) has a diverging .tm_sec, so
that on the receiver side the happening conversion to its locale
timezone does not loose the second precision.  RFC 5322 supports
the special -0000 notation to mean

  Though "-0000" also indicates Universal Time, it is used to
  indicate that the time was generated on a system that may be in
  a local time zone other than Universal Time and that the
  date-time contains no information about the local time zone

so use that in this very specific case.

  [1] https://mailarchive.ietf.org/arch/msg/emailcore/OEwxOYEabI518_0m7r9vO57O0HI/