mem: bag DBG() soil non ALLOC_ZERO regions; add MEM_CONF_FILLER_SET..
While working on the S-nail $(( )) implementation that i also want
to contribute to busybox i recognized that busybox produced an
error that just did not happen with S-nail.

Whereas the overall reason was a terrible miss that could have
been found gracefully with an according test, it was the
LOFI_ALLOC() memory content that revealed it.  That is, actually
not!  But LOFI_ALLOC() was #define'd to alloca(3) for busybox.

So for one _do_ soil _all_ memory in debug code, and furthermore
make the actually used filler configurable with a new option.
For simplicity, introduce this for su_state_create() via
su_STATE_CREATE_RANDOM_MEM_FILLER.

    u8 mf;

    su_random_builtin_generate(&mf, sizeof(mf), su_STATE_ERR_NOPASS);
    /* The patterns 0 and -1 are more likely to reveal problems */
    switch(mf & 0x3){
    case 0x0: mf = 0x00; break;
    case 0x1: mf = 0xFFu; break;
    default: break;
    }
    su_mem_set_conf(su_MEM_CONF_FILLER_SET, mf);

S-nail can now randomize the used filler upon program start:
with that (mf=0x00 to be exact) the problem would have been revealed
by our then more incomplete test with a somehow high probability.
(It would not have been revealed with the default 0xAA filler.)