[-] Handle $(( )) fallout..
Now that we fully support sh(1)ell style $(( )) evaluation,
we must be aware of constructs like

  set i=0
  if 1
    echo bla
  elif $(( i += 1 ))
    echo boo
  end
  echo $i

To be shell compatible "echo $i" must say 0, which it yet did not
because the arguments are evaluated before we even know that it is
`elif' that shall be executed next.

So change the infrastructure to look forward if the command to
evaluate is a mx_CMD_ARG_F one, and react accordingly.