a_mt_classify_round(): (hack) do encode CRLF (John Holder)..
John reported on the ML that a DOS file will loose its CRLF line
terminators in favourite of LF ones when saving the file.

This is a long story about a_folder_mbox_setptr() normalizing MBOX
to RFC 4155 (for quite some years, CRLF related comment lost with
[dc9655740abbe50628d4556d3f9d371fa0c419d3]
(a_folder_mbox_setptr(): improve/correct MBOX detection
(Dr. Werner Fink).., 2018-12-13)), etc etc.

The real healing will come with the MIME layer rewrite, which will
create a tree of objects that know about themselves and will be
dumped as desired, and eventually be re-encoded along the way, to
adhere to RFC 4155 (\n) or RFC 5322 (\r\n), etc etc.

For now we unfortunately have to enforce a MIME encoding for any
file that uses CRLF line sequences (before
[cacee1b57027a8fff3cc094e66645ce9e1dd9209] (Rewrite file-content
classification.., 2013-01-02) they seem to have been binary, but
at least did not loose their line ending), so that on the receiver
side a S-nail, or a likewise primitive software, will be able to
properly save it with its desired line-ending.

This is only about saving by the receiver, we did keep CRLF in the
data correctly, as required by the standard.  (But it is a mess!)