mx_cmd_arg_parse(): fix: misbehaviour..
Until now a final GREEDY argument that entered the jredo: loop due
to conditions (remaining input to be injected due to semicolon, or
cookie set) would fail because cad_idx would never be incremented
over it, iff the argument would not also be OPTIONal.
Instead flag that we entered the jredo: loop, and do not cause any
failure if seen.