mk/make-config.sh: write out config later..
This fixes a long standing bug in that many detections are not reflected
in the configuration; for example, OPT_USAN may not be supported by the
compiler, but nonetheless the generated files will state OPT_USAN to be
present.  This is not so terrible as it sounds as this information is
actually not used, but it is definetely not correct behaviour.

Rename option_evaluate() to option_cleanup() and fix the false claims
that shell snippets embedded in options are evaluated there; they were
in fact evaluated once option_join_rc() reread the newly created file
that contains the joined (command line, make.rc) options!

Do not use MAKEFLAGS nor COMMLINE (containing MAKEFLAGS) for "configuration
changed" detection as GNU make 4.4 started embedding volatile information
that contains random-based characters.