[startup-sani] (BWDIC!) Change again *on-mailbox-event*..
Now that we have diversified *folder-hook* to on-mailbox-open and
-newmail, change that once again to an "-event" multiplexer that
documents

  "open", "enter", "newmail", "leave", "close"

but for now only supports open, newmail, close.
Like this we for one have a complete life cycle, and could later
support multiple concurrently opened mailboxes, and have the user
write hooks which properly deal with that.

Introduce a new :N colon modifier to be able to specify only the
newest messages for the "newmail" event; remove the Heirloom way
of doing this (until now folder-hook for "newmail" would only show
the newest no matter what).