Remove *add-file-recipients*..
It was merely for testing only, generated emails will be invalid.
(It seems i added this in 2012, and noted that pipe addressees
have simply been included before that, but file ones never. Just
as bogus.)

The mechanism was then also used to display more content etc to
interactive user when editing the message draft with headers, but
this we can now do in a better way.