[SU.3] cs_dict: add data_space() / CS_DICT_DATA_SPACE_RAW "logic"..
Very often i found the need to deal with allocations etc for
simple datatypes stored in a dictionary, yet only a few bytes,
and the dict has to do all the dance for its node anyway.

So, even though it makes the dictionary logic even more
complicated and horrifying, let us add some more of it so that it
can reserve some data_space() bytes for non-DICT_OWNS instances.