(BWDIC!) n_namelist_vaporise_head(): no more *metoo*, ..that was an error..
No longer process "`alternates' (but for *posix*)" for all
receivers when sending (as opposed to replying to) mail.

However, when !*posix*, add `alternates' to the is-my-name
processing that covers $LOGNAME / *from* / *sender* / *reply-to*
that correlates to *metoo* when expanding `alias'es!

The manual change is in "nail.1: polish, II (Ashton Fagg)".