(BWDIC!) `alias' (`alternates'): be sane; add `alias -[-+]' modes..
That

  alias NAME NAME

(one member, same name) skips *metoo* handling is in all BSD
Mail's as well as in SysV10 mailx, but it is not in POSIX, and it
is questionable how sane it is.  It prevents sane implementation.

BWDIC is that

  alias - NAME

now expands depending on *metoo*.  To make that controllable

  alias -- NAME
  alias -+ NAME

always work "as-if" without/with *metoo*.

The manual change is in "nail.1: polish, II (Ashton Fagg)".