[-] Scope work (adds `our', three scope mode..; fix environ link unroll bug)..
The `localopts' hack has to vanish in the future.
Our `local' modifier is nice, but `localopts' is more, and to be
able to track changes for the current and deeper scopes, add
the `our' command modifier -- the necessary functionality is more
or less already present, even though hacky.  In the future more
and more commands shall support `our'.

Add a RAW argument type to the cmd_arg_parser -- like this we can
convert any VPUT to cmd_arg_ctx, and get rid of those terrible
global bypass indicator bits.
This renders obsolete the development-branch-only "any vput
supporter auto-supports local" (read on).

Change the devel-only local extension.
It is a bad thing to overload the `local' of a command to its
`vput', let alone the command line argument evaluation.
Instead, introduce three different scopes: "pp" for command line,
"vput" for the storage, and anoher one for the command itself.

  local pp our vput CMD res $(( i = 1 + 1))

will thus place $i in the local scope, $res in the current and
deeper ones (as via temporary `localopts on'), and have no scope
for the command itself.  Ie the same as

  local set i
  localopts on
  set res
  localopts off
  vput CMD res $((i = 1 + 1))

but shorter.

While here, change quite a number of commands to cmd_arg_ctx,
mostly in order to be able to pass all those scopes through to
where shell expression parser / cmd_arg_ctx (getmsglist() etc) is
actually used.

While here, adjust mx-test.sh to always run all tests with
$KEEP_DATA=y, ie $LOOPS_SMALL as well as $LOOPS_BIG.
And resort it a little bit.  And shrink vim fold space usage.

This fixes a bug with linked environment variables and unrolling,
which i decided not to backport.  (Ie flags of unrolled target
shall be fully restored, but putenv() shall be done one last time.)

Manual change is in "nail.1: polish, II (Ashton Fagg)