Rework mime-classification a bit (Dr. Werner Fink)..
In short: for the very first time, we do it _about_ right.
For example all the headers are now processed as a single unity,
instead of in conjunction with their MIME part content, which
i deem the right thing to do.

Most dramatic is improvement for 7-bit clean headers in
a non-US-ASCII charset-7bit setting:

  $ </dev/null mailx -:/ -Smta=test -Scharset-7bit=iso-2022-jp -s test d@a
  Subject: test
  User-Agent: mailx v14.9.24
  MIME-Version: 1.0
  Content-Type: text/plain; charset=iso-2022-jp

  $ </dev/null s-nail -:/ -Smta=test -Scharset-7bit=iso-2022-jp -s test d@a
  Subject: =?iso-2022-jp?B?dGVzdA==?=
  User-Agent: s-nail v14.9.24-574-g9ad6cbbd93
  MIME-Version: 1.0

Also ^ no body, no content-type.  For example.
We still keep "defaults" otherwise, even if not needed, since i am
afraid of some programs/versions which might fail to use the
defaults if not "explicit is better".

In addition *reply-in-same-charset* actually now tries to also use
*that* character set for headers first, too.  (I think for the
first time ever.)

I expect the v15 rewrite to keep a little bit of that, but hide it
under the hood of objects (header, part, whatever).
(Of course: normal iterations aside.)