[-] [BWDIC] Add *on-compose-embed*; obsolete *splice* hack..
Well, yay!  The machinery is still a bit "hacky", but we can now
drive a macro from within the *current* process, no more need to
start a subprocess and use complicated and messy indirection that
no normal user can truly deal with.  It now is simply

  set on-compose-embed=ocm
  define ocm {
    ~< ~/.mysig
  }

to create a sig, so keep *signature* obsoleted.