[-] (REJECTED) `write': honour !mailcap ISTXT, iactive ask for ovwr+iconv..
If a user installs MIME handlers which enforce treating a MIME
part as plain text (or HTML) we should honour that when `write'ing
them out, in that we should interactively ask for character set
conversion and file overwriting, as appropriate.

Since RFC 1524 mailcap does not offer a "save" slot, explicitly
leave those out.  ("print" cannot be used, and "copiousoutput"
dumping does also not seem right to use: better "raw", then.)

NOTE
----

This does not work out for real for now, because we do not
content-transfer-decode the MIME part content before the classify
the content to check the character set etc.
It is deemed to be to bad a hack to really do this now, but it
will be the way to go for v15.