Our address parser is as bad as i say since 2013 :-(, I..
In this first step we move around stuff, and as necessary create
or adjust some interfaces, in order to isolate all the code
necessary to exchange the parser for addresses, references etc,
in a single file, names.c.

Almost two years passed since this work was done, and many commits
have happened in the meantime, so really split this work up, and
do "the actual work" in a much, much later, equally named, second
step.