Add I(nternet)M(essage)F(ormat) parser..
Add a rather generic RFC 822/2822/5322 IMF parser.
It is meant to live in the second stage, and parse superficially
prevalidated data.  For example whitespace on follow lines or
validity of line endings is not verified.

It rather follows Postel in that it allows CFWS almost everywhere,
and is very liberal in general: this is what you want, there is an
IMF message, and you want to make sense of it.
Errors can be relaxed further.