FIX: *customhdr* <-> -S freezing bug..
*customhdr* was wrong ever since [ff8b2ba1260b], (Add -C (create
custom header), change *customhdr*, 2017-12-24).

At least in combination with -S, because the SET_PRE check for VIP
variables was used to verify *customhdr*, and in case of success
the result list was already used, whereas the FROZEN check happens
thereafter, resulting in the assignment of the variable to be
aborted.  Ie: variable had correct value, but the pre-parsed
result would still be used.

So instead give n_header_add_custom() a new mode which only
verifies, and create our result no earlier but in a new SET_POST.
The little bit of double work has to be accepted for now. XXX