[-] *on-program-exit*: define it more (Robert Elz)..
It always has been a bad idea to let it run after the terminal is
teared down.  But especially so given that it can call stuff which
may do "things".  Now lesser so as we explicitly set MB_VOID in
order to disallow access to the current folder.  (And longing for
the day these are objects which then would no longer exist.)