termios: support being a (background)& process (Robin Humble, Chet Ramey)..
Robin Humble reported

   $ ( echo blah | Mail root ) &
  [1] 2754649
   $ ^M^M^M^M^C^C

  [1]+  Stopped                 ( echo blah | Mail root )
   $ fg
  ( echo blah | Mail root )

and said

  I dug a bit deeper and it seems s-nail mistakenly tries to set
  terminal options in this use-case?

Well we did not yet support running in the background at all, and
after a ridiculuous thread i started on bug-bash@ i got a hint of
Chet Ramey pointing to later "bg"s .. so that needs to be taken
into account also.

Fix all this by simply rejecting to execute in the background, by
sending ourselves a SIGTSTP signal.
Ie, the termios code is only entered if we need to set the
terminal attributes, but that will not do when we are in the
background, and thus.
When starting up in the background we may not query terminal
attributes, as the terminal owner could have placed it in whatever
state, say.  (That was the real bug here.)

There remain optimization possibilities in that if we start
a child process we may want to restore terminal attributes
needlessly (as happened in the situation above).

We also did not document that it should be

   $ ( echo blah | Mail root >/dev/null) &

to avoid that the interactive code kicks in.
We do that still falsely after this commit, in that we look for
TTYANY instead of doing anything such *only* in INTERACTIVE mode.