mx_child_wait(): brainos..
.ce_done will never be set since child signal is blocked.
Does not really matter, but is false.

Other than that we unfortunately need to extend the range of
blocked signals beyond a memory deallocation, because, in the case
$PAGER is running, we may sit on that wait for a long time, and if
for example an interrupt happens it will stuck in the queue of
signals to be processed, only to cause a jump upon unblocking the
signals.  SA_RESTART and siglongjmp() are evil.