gcc 15.1 revealed string init fault of mine (Martin Uecker)..
The -Wunterminated-string-initialization warning of gcc 15.1
brought up a few warnings like

  /home/steffen/src/nail.git/src/mx/cmd-vexpr.c:170:50: warning: initializer-string for array of 'char' truncates NUL terminator but destination lacks 'nonstring' attribute (16 chars into 15 available) [-Wunterminated-string-initialization]
    170 |         {a_X(a_VEXPR_CMD_AGN_DATE_STAMP_UTC, 0), "date-stamp-utc\0"},

I lived under the impression that a compiler would see the
trailing NUL and accept the string is already terminated.
Like that i can easily detect char[] are sufficiently spaced,
since i get a warning if the initializer will not fit.

I found

  https://gcc.gnu.org/bugzilla/show_bug.cgi?id=116082

during my search, where i cited K&R, and that i know about the
.asciz / .ascii difference, but again, thought ...

But uecker@ proved i am wrong.  A bit shocked, but luckily
i started using this after 2010/1, so no old C++ code is infected.
So here we drop this notation.

Alongside this it was detected that the `vexpr' subcommand
date-stamp-utc never fit in the char[].  How that?  I do not know.