Refine SIG(HUP|INT) (Dag-Erling Smørgrav, Kyle Evans)..
I saw commits by the mentioned in the FreeBSD repository.
We are not POSIX compliant as that says:

  When mailx is in Send Mode and standard input is not a terminal,
  it shall take the standard action for all signals.

  In Receive Mode, or in Send Mode when standard input is a terminal,[.]
  ...
  The mailx utility shall take the standard action for all other signals.

For now it is impossible to comply to that because the entire
software is heavily based upon siglongjmp(), and the POP3, IMAP
and Maildir parts have to be rewritten completely.  The rest also,
of course, but especially the network protocols are driven by
signals.  We need(ed) to mitigate that.  Therefore this massive
protection of the go.c loops etc.  etc etc.

From Kyle Evans we take that the SIGHUP handler in go.c is not
necessary at all, as the default actions is to terminate, and
savedeadletter(), used in ancient times he says, no longer
happens.  Since the exit(2) in the handler only shadows the actual
exit cause, this is doable.  Other parts of the software still
install SIGHUP handlers etc, but reraise etc, and thus this is ok.

Dag-Erling Smørgrav started that commit trail on 2025-04-27 by not
catching signals, as above, and adding lots of tests.  We cannot
do that, as above.  We cannot test either, because i had removed
the -I option as *interactive* is not longer a variable, and being
interactive is a property of our environment: we thus need an
"expect"-alike openpty() tester to truly test interactive
sessions.  This was planned for v14.10, in conjunction with
net-test, but it yet did not happen.  (It will be hard due to
combinations of colour, termios, termcap, and MLE: lots of tests
can be written, and UI will likely change quite a bit in the
future, if i live long enough, too.)

However, what can be done in go.c and collect.c is that the
signal handlers are finally made safe in that they only jump, and
the lots of code that until now ran in there was moved to regular
execution environments.  This goes in line with the
(mx_child_wait(): brainos..) commit that started in January.