a_coll_dead_save(): added: include headers when *save* in $DEAD!..
This is still bad, of course, because we simply dump the message
as if we would *editalong*.  It is not MIME aware.
This all awaits v15, when we create a tree of objects, which then
can either be DUMP_TO_WIRE or DUMP_TO_UI: at that time we simply
serialize the tree as it is TO_WIRE; which then affects all the
many places where $DEAD is used, of course.  It always was a mess.