header_is_known(): at times only search virtuals (rahl)..
rahl reported on IRC that git-send-email does not work since Date:
comes in via git(1), but is anyway created by us.
My tries to use digmsg to remove the header failed, because
header_is_known() says we do know the header Date:, and thus we
reject removing it.

This is all borked: when an email comes in via -t, we do not
really know Date:, we treat it as any user header; and digmsg etc
should be capable to remove any such header, unless either struct
header has an explicit field (which we address directly), or it
truly is a virtual-only header, ie Mailx-Comand.
So make distinction possible.