Add `(un)?headerorder' command(s) (plus ~M+ FIX)..
While here fix a bug that is in since about 2019-09-09: in the
"fancy" quote-fold mode ~M etc may end up without actually quoting
headers because send.c:sendpart() will -- for display purposes --
unite header lines with all their follow-ups, but not append a \n
LF to that string buffer: these \n LFs are generated by writing to
the file pointer later on, but are not part of the string buffer.
Yet that string buffer is passed to the quoteflt, and since that
does not see \n LF it will not dump out the data, assuming more is
to come!!  That data is effectively lost thus.

The fix is easy: no longer generate these \n on the file pointer
directly, but simply (re!) terminate the string buffer with LF,
then pass it along the quote filter.