Honour *cols* for SEND_TODISP*, and SEND_TOQUOTE*..
A gross hack that implements header folding at the wrong place,
working on encoded headers; but we want the functionality, and we
cannot do this in the MIME layer (mx_mime_display_from_header())
because there colour etc, and much more, is not unavailable.
Despaired need for the v15 MIME rewrite.  Also here.