FIXME [-] mx_mime_enc_mustquote(): we do RFC 2047 wrong..
Yes, 2047 5.3. gives very special rules for address lists etc.
Yet, when we encode, we do not know the name of the header we
work, and therefore this never worked as it should.
For now tighten them all.

Furthermore, we should never encode within a quoted string.
Yet the state whether we are within is not available when we
work, so this never worked as it should.
It has to be noted that number one (mutt) for example simply
looses the notion of "quoted string", so we are in bad company
being so bad.  We are bad.  I hate it