Adjust mx_mime_charset_iter_reset()..
This used to be fixup commit for (Rework mime-classification a bit
(Dr. Werner Fink)[..]), but rebasing would have been a disaster.
Thus i join it with a partiall completed (mx_mime_enc_mustquote():
we do RFC 2047 wrong[..]) commit the base changes of which are yet
missing, yet nice code reordering is ok to have now.