mx_visual_info(): FIX: treat mbrtowc()==-2 just like -1..
..in case of VISUAL_INFO_SKIP_ERRORS being requested a replacement
character will be used even in "incomplete multibyte character"
error cases.

Until now we simply did not consume the input and return FAL0
regardless of VISUAL_INFO_SKIP_ERRORS, yet certain callers did not
handle the return value in conjunction with that bit.
I got an endless loop via

  set d$'\xE4'=; set

and the reason was that the shell quoting code did not test the
return value of mx_visual_info(), and then decrementing a length
by 0 because there was actually no output means that the same
buffer and length will be worked on, forever.
The error is already in v14.9.25, but hard to reach.